Split test helper functions
